requirements.txt
jira
pandas
requests
pyyaml
openai
anthropic
python-dotenv
boto3
botocore
